Relocate context
